backend/requirements/base.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==1.3.1
fastapi==0.95.2
fastapi-crudrouter==0.8.6
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
pydantic==1.10.13
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.5.2
SQLAlchemy==2.0.23
starlette==0.27.0
typer==0.9.0
typing-inspection==0.4.1
typing-extensions>=4.12.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
